테스트 케이스 및 버그 수정

+++ src/main/domain/ledger/category/category/service/category_service.py
@@ -1,7 +1,7 @@
 from fastapi import HTTPException
 from sqlalchemy.orm import Session
 
-from domain.ledger.category.category.dto import CreateCategoryDTO
+from domain.ledger.category.category.dto import CreateCategoryDTO, DeleteCategoryParams
 from domain.ledger.category.category.dto.category_response_dto import CategoryResponseDto
 from domain.ledger.category.category.dto.edit_category_dto import EditCategoryDto
 from domain.ledger.category.category.dto.search_category_params import SearchCategoryParams
@@ -11,6 +11,8 @@ from domain.ledger.category.item.dto import CreateItemDto
 from domain.ledger.category.item.dto.item_response_dto import ItemResponseDto
 from domain.ledger.category.item.entity import Item
 from domain.ledger.category.item.repository import ItemRepository
+from domain.ledger.receipt.entity import Receipt
+from domain.organization.organization.repository import OrganizationRepository
 
 
 class CategoryService:
@@ -19,11 +21,18 @@ class CategoryService:
             self,
             category_repository: CategoryRepository,
             item_repository: ItemRepository,
+            organization_repository: OrganizationRepository,
     ):
         self.category_repository = category_repository
         self.item_repository = item_repository
+        self.organization_repository = organization_repository
 
     async def create(self,db:Session, create_category: CreateCategoryDTO):
+        organization = await self.organization_repository.find_by_id(db, create_category.organization_id)
+        if not organization:
+            raise HTTPException(status_code=404, detail="Organization not found")
+        if not organization.start_year <= create_category.year <= organization.end_year:
+            raise HTTPException(status_code=400, detail="Invalid year")
         category:Category = await self.category_repository.create_category(
             db=db,
             create_category_dto=create_category
@@ -56,13 +65,16 @@ class CategoryService:
         return category_dtos
 
     async def update(self, db:Session, edit_category_dto:EditCategoryDto):
-        category = await self.category_repository.find_category_by_id(db, edit_category_dto.category_id)
+        category = await self.category_repository.find_by_organization_and_id(db, edit_category_dto.organization_id, edit_category_dto.category_id)
         if not category:
             raise HTTPException(status_code=404, detail="Category not found")
         await self.category_repository.update_category(db, category, edit_category_dto.category_name)
 
-    async def delete(self, db:Session, category_id:int):
-        category = await self.category_repository.find_category_by_id(db, category_id)
+    async def delete(self, db:Session, delete_category:DeleteCategoryParams):
+        category:Category = await self.category_repository.find_by_organization_and_id(db, delete_category.organization_id, delete_category.category_id)
         if not category:
             raise HTTPException(status_code=404, detail="Category not found")
+        receipts:list[Receipt] = category.receipts
+        if len(receipts) != 0:
+            raise HTTPException(status_code=400, detail="Category has receipts")
         await self.category_repository.delete(db, category)
